feature: changed method of printing stacktrace to log, moved ErrorResponse to dto package
