+++ src/main/java/ru/yandex/practicum/filmorate/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.model;
+package ru.yandex.practicum.filmorate.dto;
 
 import lombok.Value;
